refactor : profile 에 따른 서비스 및 엘라스틱 연결에 관련한 코드들 간소화
