+++ prography-search/src/main/kotlin/org/prography/search/config/property/ElasticsearchProperty.kt
@@ -1,8 +1,11 @@
 package org.prography.search.config.property
 
+import org.springframework.boot.context.properties.ConfigurationProperties
+
 /**
  * Elasticsearch 관련 설정 값
  */
+@ConfigurationProperties(prefix = "spring.data.elasticsearch")
 data class ElasticsearchProperty(
     /**
      * 연결 HOST 주소 e.g localhost, 127.0.0.1
